// frontend/src/lib/services/model/notificationItem.ts
/**
 * Bankatek API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 */


export interface NotificationItem { 
    id?: string;
    title?: string;
    text?: string;
    kind?: string;
    date?: string;
    read?: boolean;
    path?: string;
}